backend/requirements.txt
Django==4.2.13
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
python-dotenv==1.0.1
psycopg2-binary==2.9.10
supabase==2.4.4
requests==2.32.3
PyJWT==2.8.1
cryptography==42.0.8
